Bug 21629 - equivalent DIEs must be of the same DIE source
A prerequisite to represent the relationship "DIE 'D' has canonical
DIE C", is that D and C come from the same DIE source.  A DIE source
being either the main debug info file, the alternate debug info file
(both having DIEs in their .debug_info section), or a .debug_type
section.

This prerequisite is a result of the design of the current
libabigail's DWARF reader.

Unfortunately the code that compares DIEs and that updates the
canonical DIEs of the compared DIEs (in case the two DIEs are
equivalent) sometimes fails to honour that prerequisite, causing other
troubles down the road.

This patch fixes that by ensuring that equivalent DIEs must be of the
same DIE source.

The patch doesn't embed a regression test as this problem was
discovered using the
https://pagure.io/libabigail-selfcheck/blob/master/f/selfcheck.py
testing framework

	* src/abg-dwarf-reader.cc (compare_dies): Don't propagate
	a canonical DIE to a DIE that comes from a different source.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>