Don't add empty translation unit to corpus
In debug info, some translation units are known to be empty because
they don't carry any DIE.  It's thus useless to add those translation
units to the ABI corpus.

Note that in some packages like gcc and glibc (in Fedora 25) the same
empty translation unit can be present in the debug info more than
once.  This violates a Libabigail invariant that asserts that a given
translation unit (designated by its path) shall be present only once
in the debug info.

	* src/abg-dwarf-reader.cc (build_translation_unit_and_add_to_ir):
	A translation unit DIE that has no child DIE shall not be added to
	the current ABI corpus.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>