Avoid comparing kernel.img file from the grub2 package
The command "fedabipkgdiff --self-compare --from fc25 grub2" returns
with an error because no debug info can be found for the file
kernel.img.  That file doesn't have any debug info shipped for it.
And we shouldn't try to compare it anyway.

This patch updates the default suppression file shipped by libabigail
to make it avoid compare kernel.img files.

	* default.abignore: Do not compare kernel.img files.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>