speed up class type lookup in a corpus
When a class type lookup fails in a the type map of a corpus, the code
goes look the class up in the type maps of a every single translation
unit.

This design dates back from the time where there could be only one
type of a given name present in in the type map of a corpus.  When
several types with the same were found, then each type would be
present in the type map of its translation unit, but then no type with
that name would be present in the type map of the corpus.

As a result, when a type was not found in the type map of the corpus,
we were looking for it in the type maps of the translation units.

Now that the type map of the corpus can carry several types with the
same name, we don't need to look in the translation unit anymore.

This patch implements that adjustment.

	* src/abg-ir.cc (lookup_class_type): In the overload that looks
	for a class name denoted by an interned_string in the corpus, do
	not look for the class in the translation units when the type
	wasn't found in the type map of the corpus.
	(maybe_update_types_lookup_map): Remove the
	erase_if_exists_already parameter and the code that uses it.
	(lookup_class_type_through_translation_units): Remove this
	function that is now useless.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>