Don't crash on classes that differ in their virtual member fn count
When comparing classes that differ in their number of virtual member
functions, it can happen that we crash because we should have got out
earlier.

This patch fixes that.  That problem happens on the test provided in
PR libabigail/17948 -- which is too big to be included in the test
suite.

	* src/abg-ir.cc (equals): In the overload for class_decl, when we
	detect that the virtual member function counts are different, get
	out, even when we are being asked about the kind of the change.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>