Use an unordered map for canonical DIE offsets
Until now, the canonical DIE offsets map (the map that associates a
canonical DIE offset to a given DIE offset) was implemented in the
DWARF reader by using vectors.  This proves to use a lot of memory
(more than 25GB!!) for huge vmlinux binaries.  So much that the thing
was too slow (hey, yeah, because it was swapping out!) on vmlinux
kernel binary of 3GB of size, just to abidw it out.

This patch changes that to use an unordered map for this.  It now uses
~ 4GB of peak RAM memory, so I can actually abidw the vmlinux of 3GB
of size on my laptop.  It's still taking 16 minutes (!!) to complete,
but at least it's progress, as it's completed at least, without
swapping out.

And it doesn't see to noticeably slowdown the test suite.

	* src/abg-dwarf-reader.cc
	(read_context::canonical_type_die_offsets_): Renamed the
	canonical_type_die_vecs_ data member into this.
	(read_context::canonical_decl_die_offsets_): Renamed the
	canonical_decl_die_vecs_ data member into this.
	(read_context::{initialize, compute_canonical_die_offset,
	compute_canonical_die, get_canonical_die,
	get_or_compute_canonical_die, set_canonical_die_offset,
	get_canonical_die_offset}): Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>