[abixml writer] Store pointers to emitted types rather than type-ids
To record the emitted types, the abixml writer records the emitted
type *ID*s.  So to lookup an emitted type, it needs to lookup the ID
of the type first, and then lookup that type-id in the set of emitted
type IDs.  We are doing twice as much work as we should and profiling
the writting of a big abixml showed that it's quite taxing.

This patch makes the write records the set of emitted types directly.
No need to go through the emitted type *ID*s anymore.  So we save one
map lookup.  This incurs a 16% speedup when writting an abixml file
for a big (3GB) vmlinux file.

	* src/abg-writer.cc (type_ptr_set_type): Declare new typedef.
	(writer_context::m_emitted_type_id_map): Remove this data member.
	(writer_context::m_emitted_type_set): Add a new data member.
	(writer_context::{record_type_id_as_emitted, type_id_is_emitted,
	clear_emitted_types_map}): Remove these member functions.
	(writer_context::{record_type_as_emitted, type_is_emitted}): Use
	the new m_emitted_type_set data member above.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>