Don't make system headers depend on config.h
While building on an 32 bit intel platform, compilation of
abipkgdiff.cc failed because  fts.h wouldn't compile with
the macro _FILE_OFFSET_BITS set to 64.  That macro is set by config.h!

The idea of this patch is to make config.h have an effect only on
libabigail's own header files, not on system header files.

	* tools/abipkgdiff.cc: Include config.h after system headers and
	before libabigail's headers.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>