Bug 22076 - Disable fedabipkgdiff for old koji clients
On some platforms, the koji client is so old that the koji object
doesn't have the .read_config method that we need.

Unfortunately, the koji client library module doesn't come with a
__version__ variable that carries its version number.  So we cannot
use the version of the library.  This patch thus tries to invoke the
koji.read_config method and see if it works.  If it does, the koji
client is deemed recent enough.  Otherwise the fedabipkgdiff feature
is disabled.

	* configure.ac: Try to invoke the koji.read_config method.  If it
	fails then disable the fedabipkgdiff feature.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>