Bug 22437 - Make fedabipkgdiff use all debug info RPMs of a sub-RPM
On ppc64, at least, it can happen that a devel RPM foo-devel.rpm needs
not only the foo-devel-debuginfo.rpm but also the foo-debuginfo.rpm.
In other words, the devel sub-package needs all the debug info RPMs of
foo.rpm.

To make things general, this patch modifies fedabipkgdiff so that all
debuginfo packages are taken into account when looking at a devel
package.  This is done on all architectures, not just ppc64.

	* tools/fedabipkgdiff (RPM::get_all_debuginfo_rpms): Define new
	member function.
	(RPM::generate_comparison_halves): The ancillary debuginfo RPM of
	a given RPM now has a list type; there can be more than one
	debuginfo RPM associated to a given RPM, especially if the RPM is
	a devel one.
	(format_debug_info_pkg_options): Define new function.
	(abipkgdiff): Use the new function above.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>