Only consider local changes when filtering subtype changes
In the leaf reporter, when counting subtype changes of struct/unions,
we forget that we should be looking at *local* changes (filtered or
not) only.  This leads to some weird change reports like this, when
comparing RHEL 7.0 and RHEL 7.1 kernels:

    'struct pmu' changed:
      type size changed from 1408 to 1536 bits
      2 data member insertions:
	'module* pmu::module', at offset 1408 (in bits) at perf_event.h:282:1
	'int pmu::capabilities', at offset 1472 (in bits) at perf_event.h:287:1
      there are data member changes:

In the report above, you see that the reporter thinks that there are
further changes to be reported, but then none of them are actually
reported.

This is due to the fact that when the reporter determines if there are
data member changes or not, it considers all changes, not just local
changes.  But then when it's time to actually report the changes, then
it only considers local changes.  So there is a difference in there.

This patch makes the leaf reporter to consider local changes only,
when it determines if there are data member changes.

	* src/abg-comparison-priv.h
	(class_or_union_diff::priv::{count_filtered_subtype_changed_dm,
	count_filtered_changed_dm}): Take an additional flag.
	* src/abg-comparison.cc
	(class_or_union_diff::priv::{count_filtered_subtype_changed_dm,
	count_filtered_changed_dm}): Likewise.  When asked, only consider
	local changes.
	* src/abg-leaf-reporter.cc (leaf_reporter::report): Consider
	reporting only the *net local* data member changes.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>