Don't possibly forget type definition when reading a CorpusGroup
Suppose type T was declared in the main corpus of a group, and suppose
it's a declaration-only type.  Later, when loading another corpus of
the group, we see a definition of T.  We should load the definition of
T and not just say that we have already seen T from the main corpus
and we just keep its declaration and never get its definition.

This is what this patch does.

	* src/abg-dwarf-reader.cc (add_or_update_class_type): Look for
	declaration-only-ness to determine if we've already seen the same
	type from the main corpus of the group.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>