Improve detection of local *type* changes
For variables and data member, we are quite gross in the way we detect
a local type change.  Basically, if a textual representation of the
variable (and thus of its type) changes, then we consider that the
variable has a local (possibly type) change.

This leads us to (wrongly) report changes like this:

'struct S1 at test-44-anonymous-data-member-v0.c:1:1' changed:
  type size hasn't changed
  there are data member changes:
   anonymous data member at offset 32 (in bits) changed from:
     union {int b; float c;}
   to:
     union {int b; float c; char e;}

Here, you see that the textual representation of the anonymous data
member (of union type) changed from:

    union {int b; float c;}
to:
    union {int b; float c; char e;}

You see that although the textual representation of the type changed,
the *structure* of the type hasn't really changed.  I am using the
"vague" term structure, on purpose.  Here, in the case of a union, the
structure hasn't change because the size of the union hasn't changed.

This patch thus introduces the concept of "similarity of type
structures".  That is, even if two types are different, if "their
structure is similar", then the type change is not a local type
change.

More precisely, here is what we mean by type similarity:

    Two indirect types (pointers, references) have similar structure
    if their underlying types are of the same kind and have the same
    name.  In this indirect types case, the size of the underlying
    type does not matter.

    Two direct types (i.e, non indirect) have a similar structure if
    they have the same kind, name and size.  Two class types have
    similar structure if they have the same name, size, and if their
    data members have similar types.

This patch then uses that similarity concept to detect local type
changes.

	* include/abg-fwd.h (is_type_decl): Declare new overload for
	type_base*.
	(types_have_similar_structure): Declare new function.
	* src/abg-comparison.cc
	(class_or_union_diff::priv::count_filtered_changed_dm): Even when
	looking at local changes only, do not forget to count nodes that
	were filtered out.
	* src/abg-ir.cc (types_have_similar_structure): Define new
	function.
	(is_type_decl): Define new overload for
	type_base*.
	(is_enum_type):
	(equals): In the overload for var_decl, use the new
	types_have_similar_structure function to detect local (type)
	changes.
	* src/abg-reporter-priv.cc (represent): In the overload for
	var_decl, use the diff::has_local_changes function to detect local
	changes, now that we can better detect local changes.
	* tests/data/test-diff-filter/test44-anonymous-data-member-report-1.txt:
	Adjust.
	* tests/data/test-diff-suppr/test36-leaf-report-0.txt: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>